refactor: split unit test to file
