+++ crates/bindizr-dns/src/nsupdate/auth/mod.rs
@@ -204,16 +204,4 @@ fn build_tsig_signed_data(query_data: &[u8], tsig: &TsigRecord) -> Result<Vec<u8
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn encode_canonical_name_lowercases_key_name() {
-        assert_eq!(
-            encode_canonical_name("Nsupdate-Key.").unwrap(),
-            vec![
-                12, b'n', b's', b'u', b'p', b'd', b'a', b't', b'e', b'-', b'k', b'e', b'y', 0,
-            ]
-        );
-    }
-}
+mod tests;
